requirements.txt
# Core SDK and dependencies
claude-agent-sdk>=0.1.10,<0.2.0
websockets>=15.0.1,<16.0.0
python-dotenv>=1.2.1,<2.0.0

# HTTP clients
requests>=2.31.0,<3.0.0
aiohttp>=3.13.2,<4.0.0
httpx>=0.24.0,<1.0.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38.0,<2.0.0
google-auth>=2.23.0,<3.0.0

# Security
bcrypt>=4.0.0,<5.0.0
PyJWT>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.11.0,<4.0.0

# Context Orchestrator dependencies
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
pygithub>=2.1.0,<3.0.0
anthropic>=0.18.0,<1.0.0

# Template rendering
jinja2>=3.1.0,<4.0.0
